perf: make website bundle with less bloat
Websites, such as a payment checkout page, take very unreasonable
amount of time and network resources to load.

This is because frappe-web.bundle does not differentiate at all.

However, a guest visitor, unlike a regular user, won't have the
heavy stuff cached.

This PR bundles a lightwight website bundle that can be used
in, again for example, payment checkout pages or other views.